moved service map reset button to service map modal reset closes the modal
